requirements.txt
langchain_community
langchain
langchain-text-splitters
unstructured
langchain-openai
docx2txt
faiss-cpu
langchain-tools
cassio
# gmail libraries
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
# calendar libraries
langchain-google-calendar-tools
langchain-google-community
tzdata
pytz
# fast api
fastapi
uvicorn
# flask 
flask
